LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/input-type-change-value-expected.txt



FAIL Without focus assert_equals: Invalid color should return a non-empty sanitized value expected "#000000" but got ""
FAIL With focus assert_equals: Invalid color should return a non-empty sanitized value expected "#000000" but got ""
